restructured db connections
