+++ app/data_layer/database/crud/sqlite/websocket_crud.py
@@ -8,8 +8,8 @@ from typing import Generator, cast
 from sqlalchemy.dialects.sqlite import insert
 from sqlmodel import Session, select
 
+from app.data_layer.database.db_connections.sqlite import get_session
 from app.data_layer.database.models.websocket_model import SocketStockPriceInfo
-from app.data_layer.database.crud.sqlite_db_connection import get_session
 from app.utils.common.logger import get_logger
 
 logger = get_logger(Path(__file__).name)
