+++ app/data_layer/database/db_connections/sqlite.py
@@ -8,8 +8,8 @@ from typing import Generator
 
 from sqlalchemy.engine import Engine
 from sqlmodel import Session, SQLModel, create_engine
-from app.data_layer.database.models import *
 
+import app.data_layer.database.models  # pylint: disable=unused-import
 from app.utils.urls import SQLITE_DB_URL
 
 sqlite_engine = create_engine(SQLITE_DB_URL)
