+++ tests/data_layer/data_saver/test_sqlite_saver.py
@@ -11,8 +11,8 @@ from pytest_mock import MockerFixture, MockType
 
 from app.data_layer.data_saver import DataSaver, SqliteDataSaver
 from app.data_layer.database.crud.sqlite.websocket_crud import get_all_stock_price_info
+from app.data_layer.database.db_connections.sqlite import get_session
 from app.data_layer.database.models.websocket_model import SocketStockPriceInfo
-from app.data_layer.database.crud.sqlite_db_connection import get_session
 from app.utils.common import init_from_cfg
 
 Message = namedtuple("Message", ["value"])
