+++ tests/data_layer/database/sqlite/crud/test_smartapi_crud.py
@@ -10,11 +10,8 @@ from app.data_layer.database.crud.sqlite.smartapi_crud import (
     get_smartapi_tokens_by_all_conditions,
     get_smartapi_tokens_by_any_condition,
 )
+from app.data_layer.database.db_connections.sqlite import get_session, sqlite_engine
 from app.data_layer.database.models.smartapi_model import SmartAPIToken
-from app.data_layer.database.crud.sqlite_db_connection import (
-    get_session,
-    sqlite_engine,
-)
 from app.utils.startup_utils import create_smartapi_tokens_db
 from app.utils.urls import SQLITE_DB_URL
